AMP-14056
git-svn-id: file:///svn/branches/AMP_2_3_RELEASE@19349 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/permissionmanager/components/features/fields/AmpPMPagingNavigation.java
@@ -46,12 +46,12 @@ public class AmpPMPagingNavigation extends AjaxPagingNavigation {
 	@Override
 	protected void populateItem(LoopItem loopItem) {
 		// Get the index of page this link shall point to
-		final int pageIndex = loopItem.getIndex();
+		final long pageIndex = loopItem.getIndex();
 		
 		// Add a page link pointing to the page
 		final AbstractLink link = super.newPagingNavigationLink("pageLink", pageable, pageIndex);
 		link.add(new TitleAppender(pageIndex));
-		int currentPage = pageable.getCurrentPage();
+		long currentPage = pageable.getCurrentPage();
 		if(currentPage == pageIndex)
 			link.add(new AttributeModifier("class",new Model("paging_sel")));
 		loopItem.add(link);
@@ -75,7 +75,7 @@ public class AmpPMPagingNavigation extends AjaxPagingNavigation {
 		/** resource key for the message */
 		private static final String RES = "PagingNavigation.page";
 		/** page number */
-		private final int page;
+		private final long page;
 
 		/**
 		 * Constructor
@@ -83,7 +83,7 @@ public class AmpPMPagingNavigation extends AjaxPagingNavigation {
 		 * @param page
 		 *            page number to use as the ${page} var
 		 */
-		public TitleAppender(int page)
+		public TitleAppender(long page)
 		{
 			this.page = page;
 		}
